rename test_support to fixtures
